schematic shows the paper tape reader circuit board
The schematic shows nine photo transistors for eight data channels plus one for the feed hole. Each photo transistor is shown twice (for example as CR6 and XCR6) and the PCB layout shows two footprints. That was to allow two different photo transistor packages to be used. A total of nine photo transistors assembled in the final board. The extra footprints would be removed if the schematic and PCB are ever revised.